add(genre): implement user decorator & validation

+++ src/genre/genre.service.ts
@@ -1,9 +1,13 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
+import { generateKey } from 'crypto';
 import { PrismaService } from 'src/prisma/prisma.service';
+import { User } from 'src/user/entities/user.entity';
 import { handleError } from 'src/utils/handle-error.util';
 import { CreateGenreDto } from './dto/create-genre.dto';
 import { UpdateGenreDto } from './dto/update-genre.dto';
 import { Genre } from './entities/genre.entity';
+
+
 
 @Injectable()
 export class GenreService {
@@ -23,26 +27,35 @@ export class GenreService {
     return record;
   }
 
-  findOne(id: string): Promise<Genre> {
-    return this.findById(id);
+  async create(user: User, dto: CreateGenreDto ): Promise<Genre> {
+     if(user.isAdmin){
+       const genre: Genre = { ...dto};
+       return await this.prisma.genre.create({ data: genre})
+     }else{
+      throw new UnauthorizedException('Usuário não autorizado. Contate o Administrador!')
+     }
   }
 
-  create(dto: CreateGenreDto): Promise<Genre> {
-    const data: Genre = { ...dto };
-
-    return this.prisma.genre.create({ data }).catch(handleError);
+  async update( id: string, dto: UpdateGenreDto, user: User,): Promise<Genre> {
+    if(user.isAdmin){
+      await this.findById(id);
+      const data: Partial<Genre> = { ...dto };
+      return this.prisma.genre
+      .update({
+        where: { id },
+         data
+        }).catch(handleError);
+    }else{
+      throw new UnauthorizedException('Usuário não autorizado. Contate o Administrador!')
+    }
   }
 
-  async update(id: string, dto: UpdateGenreDto): Promise<Genre> {
-    await this.findById(id);
-
-    const data: Partial<Genre> = { ...dto };
-
-    return this.prisma.genre.update({ where: { id }, data });
-  }
-
-  async delete(id: string) {
-    await this.findById(id);
-    await this.prisma.genre.delete({ where: { id } });
+  async delete(user: User, id: string) {
+    if(user.isAdmin){
+      await this.findById(id);
+      await this.prisma.genre.delete({ where: { id } });
+    }else{
+     throw new UnauthorizedException('Usuário não autorizado. Contate o Administrador!')
+    }
   }
 }
